SBERDOMA-20 fix translations, add Ticket domain module at utils/clientSchema
