+++ apps/condo/containers/TicketForm/UpdateTicketForm.tsx
@@ -4,7 +4,7 @@ import { Typography } from 'antd'
 import { useIntl } from '@core/next/intl'
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore
-import { useObject, useUpdate, convertGQLItemToFormState } from '../../utils/clientSchema/Ticket'
+import { Ticket } from '../../utils/clientSchema/Ticket'
 import { BaseTicketForm } from '../../components/BaseTicketForm'
 
 interface IUpdateTicketForm {
@@ -16,9 +16,9 @@ export const UpdateTicketForm:React.FunctionComponent<IUpdateTicketForm> = ({ id
     const LoadingMessage = intl.formatMessage({ id: 'Loading' })
 
     const { organization } = useOrganization()
-    const { obj, refetch, loading, error } = useObject({ where: { id } }, false)
+    const { obj, refetch, loading, error } = Ticket.useObject({ where: { id } }, false)
 
-    const action = useUpdate({}, refetch)
+    const action = Ticket.useUpdate({}, refetch)
     const updateAction = (value) => action(value, obj)
 
     if (error || loading) {
@@ -33,7 +33,7 @@ export const UpdateTicketForm:React.FunctionComponent<IUpdateTicketForm> = ({ id
     return (
         <BaseTicketForm
             action={updateAction}
-            initialValues={convertGQLItemToFormState(obj)}
+            initialValues={Ticket.convertGQLItemToFormState(obj)}
             organization={organization}
         />
     )
