+++ apps/condo/lang/ru.json
@@ -18,11 +18,11 @@
   "pages.condo.property.field.Organization": "Организация",
   "pages.condo.property.type.building": "Многоквартирный дом",
   "pages.condo.property.type.village": "Коттеджный поселок",
-  "pages.condo.ticket.index.PageTitle": "Задачи",
-  "pages.condo.ticket.id.PageTitle": "Задача",
-  "pages.condo.ticket.index.CreateTicketModalTitle": "Добавить задачу",
-  "pages.condo.ticket.index.EditTicketModalTitle": "Редактировать задачу",
-  "pages.condo.ticket.index.CreateTicketButtonLabel": "Добавить задачу",
+  "pages.condo.ticket.index.PageTitle": "Заявки",
+  "pages.condo.ticket.id.PageTitle": "Заявка",
+  "pages.condo.ticket.index.CreateTicketModalTitle": "Добавить заявку",
+  "pages.condo.ticket.index.EditTicketModalTitle": "Редактировать заявку",
+  "pages.condo.ticket.index.CreateTicketButtonLabel": "Добавить заявку",
   "pages.condo.ticket.field.Number": "Номер",
   "pages.condo.ticket.field.Source": "Источник",
   "pages.condo.ticket.field.Type": "Тип заявки",
@@ -34,13 +34,13 @@
   "field.FullName": "ФИО",
   "field.EMail": "E-mail",
   "field.Executor": "Исполнитель",
-  "field.Executor.description": "Исполнитель — это сотрудник, который будет устранять проблему, описанную в задаче",
+  "field.Executor.description": "Исполнитель — это сотрудник, который будет устранять проблему, описанную в заявке",
   "field.Responsible": "Ответственный",
   "field.Responsible.description": "Ответственный — контролирует работу Исполнителя",
   "pages.condo.ticket.field.Responsible": "Ответственный",
   "pages.condo.ticket.field.Description": "Суть проблемы в свободной форме",
   "pages.condo.ticket.title.UserInfo": "Информация о жителе",
-  "pages.condo.ticket.title.TicketInfo": "Информация о задаче",
+  "pages.condo.ticket.title.TicketInfo": "Информация о заявке",
   "pages.condo.ticket.field.Property": "Объект",
   "pages.condo.residents.PageTitle": "Жители",
   "pages.condo.employee.PageTitle": "Сотрудники",
@@ -48,7 +48,7 @@
   "menu.Home": "Дашборд",
   "menu.Users": "Пользователи",
   "menu.Property": "Объкты",
-  "menu.Tickets": "Задачи",
+  "menu.Tickets": "Заявки",
   "menu.Employee": "Сотрудники",
   "menu.Residents": "Жители",
   "Welcome": "Привет, {name}!",
